+++ src/components/SubToolMenu.vue
@@ -26,7 +26,7 @@ interface Color {
 const { mode } = storeToRefs(useStoreMode())
 const { setLineColor } = useStoreLine()
 const { setTextOptionValue, setTextColor } = useStoreText()
-const { setDragUrl } = useStoreImage()
+const { setDragImageUrlAndId } = useStoreImage()
 const { isLoadingImages, uploadedImages } = storeToRefs(useStoreUserImage())
 const { addImageList, deleteImageFromToolbar } = useStoreUserImage()
 
@@ -233,7 +233,7 @@ div(v-else-if="mode === 'image'" class="flex justify-center items-center bg-gray
       //- image list
       div(v-for="image of uploadedImages" :key="image.id" class="relative flex justify-center items-center")
         button(type="button" class="flex justify-center items-center absolute top-0 right-0 w-9 h-9 rounded-full bg-slate-200 hover:bg-slate-300 m-1" @click="async () => {await deleteImageFromToolbar(image);}") ✕
-        img(:id="image.id" :src="image.storageURL" class="w-full aspect-auto col-span-1 p-2 hover:cursor-grab active:cursor-grabbing" @dragstart="(e) => {setDragUrl(e);}")
+        img(:id="image.id" :src="image.storageURL" class="w-full aspect-auto col-span-1 p-2 hover:cursor-grab active:cursor-grabbing" @dragstart="(e) => {setDragImageUrlAndId(e);}")
 </template>
 
 <style scoped>
