+++ src/stores/konva/transformer.ts
@@ -7,6 +7,7 @@ import useStoreText from '@/stores/konva/text'
 import useStoreImage from '@/stores/konva/image'
 // eslint-disable-next-line import/no-cycle
 import useStoreStage from '@/stores/konva/stage'
+import useStoreUserImage from '@/stores/userImage'
 
 const useStoreTransformer = defineStore({
   id: 'transformer',
@@ -344,7 +345,7 @@ const useStoreTransformer = defineStore({
         // image
         else if (selectedNode.name() === 'image') {
           const { konvaImages } = storeToRefs(useStoreImage())
-          const { deleteUploadedImageCanvases } = useStoreImage()
+          const { deleteUploadedImageCanvases } = useStoreUserImage()
           // firestoreで画像の使用状況を更新（削除)
           if (typeof canvasId === 'string') {
             const selectedImage = konvaImages.value.find(
